refactor entry api
