+++ src/http_frontend/entry_api.rs
@@ -3,6 +3,11 @@
 //    License, v. 2.0. If a copy of the MPL was not distributed with this
 //    file, You can obtain one at https://mozilla.org/MPL/2.0/.
 
+mod query;
+mod read_batched;
+mod read_single;
+mod write;
+
 use axum::async_trait;
 use axum::body::StreamBody;
 use axum::extract::{BodyStream, FromRequest, Path, Query, State};
@@ -19,6 +24,7 @@ use crate::auth::policy::{ReadAccessPolicy, WriteAccessPolicy};
 use axum::headers;
 use axum::headers::{Expect, Header, HeaderMapExt, HeaderValue};
 
+use axum::routing::{get, head, post};
 use log::{debug, error};
 use std::pin::Pin;
 use std::str::FromStr;
@@ -27,6 +33,9 @@ use std::task::{Context, Poll};
 use std::time::Duration;
 
 use crate::core::status::{HttpError, HttpStatus};
+use crate::http_frontend::entry_api::read_batched::read_batched_records;
+use crate::http_frontend::entry_api::read_single::read_single_record;
+use crate::http_frontend::entry_api::write::write_record;
 use crate::http_frontend::middleware::check_permissions;
 use crate::http_frontend::HttpServerState;
 use crate::storage::bucket::Bucket;
@@ -107,213 +116,6 @@ fn check_and_extract_ts_or_query_id(
     Ok((query_id, ts))
 }
 
-fn fetch_and_response_batched_records(
-    bucket: &mut Bucket,
-    entry_name: &str,
-    query_id: u64,
-    empty_body: bool,
-) -> Result<impl IntoResponse, HttpError> {
-    const MAX_HEADER_SIZE: u64 = 6_000; // many http servers have a default limit of 8kb
-    const MAX_BODY_SIZE: u64 = 16_000_000; // 16mb just not to be too big
-    const MAX_RECORDS: usize = 85; // some clients have a limit of 100 headers
-
-    let make_header = |reader: &RecordReader| {
-        let name = HeaderName::from_str(&format!("x-reduct-time-{}", reader.timestamp())).unwrap();
-
-        let mut meta_data = vec![
-            reader.content_length().to_string(),
-            reader.content_type().clone(),
-        ];
-
-        let mut labels: Vec<String> = reader
-            .labels()
-            .iter()
-            .map(|(k, v)| {
-                if v.contains(",") {
-                    format!("{}=\"{}\"", k, v)
-                } else {
-                    format!("{}={}", k, v)
-                }
-            })
-            .collect();
-        labels.sort();
-
-        meta_data.append(&mut labels);
-        let value: HeaderValue = meta_data.join(",").parse().unwrap();
-
-        (name, value)
-    };
-
-    let mut header_size = 0u64;
-    let mut body_size = 0u64;
-    let mut headers = HeaderMap::new();
-    let mut readers = Vec::new();
-    let mut last = false;
-    loop {
-        let _reader = match bucket.next(entry_name, query_id) {
-            Ok((reader, _)) => {
-                {
-                    let reader_lock = reader.read().unwrap();
-
-                    let (name, value) = make_header(&reader_lock);
-                    header_size += (name.as_str().len() + value.to_str().unwrap().len() + 2) as u64;
-                    body_size += reader_lock.content_length();
-                    headers.insert(name, value);
-                }
-                readers.push(reader);
-
-                if header_size > MAX_HEADER_SIZE
-                    || body_size > MAX_BODY_SIZE
-                    || readers.len() > MAX_RECORDS
-                {
-                    // This is not correct, because we should check sizes before adding the record
-                    // but we can't know the size in advance and after next() we can't go back
-                    break;
-                }
-            }
-            Err(err) => {
-                if readers.is_empty() {
-                    return Err(err);
-                } else {
-                    if err.status() == HttpStatus::NoContent as i32 {
-                        last = true;
-                        break;
-                    } else {
-                        return Err(err);
-                    }
-                }
-            }
-        };
-    }
-
-    headers.insert("content-length", body_size.to_string().parse().unwrap());
-    headers.insert("content-type", "application/octet-stream".parse().unwrap());
-    headers.insert("x-reduct-last", last.to_string().parse().unwrap());
-
-    struct ReadersWrapper {
-        readers: Vec<Arc<RwLock<RecordReader>>>,
-        empty_body: bool,
-    }
-
-    impl Stream for ReadersWrapper {
-        type Item = Result<Bytes, HttpError>;
-
-        fn poll_next(
-            mut self: Pin<&mut ReadersWrapper>,
-            _cx: &mut Context<'_>,
-        ) -> Poll<Option<Self::Item>> {
-            if self.empty_body {
-                return Poll::Ready(None);
-            }
-
-            if self.readers.is_empty() {
-                return Poll::Ready(None);
-            }
-
-            if self.readers[0].read().unwrap().is_done() {
-                self.readers.remove(0);
-            }
-
-            match self.readers[0].write().unwrap().read() {
-                Ok(chunk) => Poll::Ready(Some(Ok(chunk.unwrap()))),
-                Err(e) => Poll::Ready(Some(Err(e))),
-            }
-        }
-        fn size_hint(&self) -> (usize, Option<usize>) {
-            (0, None)
-        }
-    }
-
-    Ok((
-        headers,
-        StreamBody::new(ReadersWrapper {
-            readers,
-            empty_body,
-        }),
-    ))
-}
-
-fn fetch_and_response_single_record(
-    bucket: &mut Bucket,
-    entry_name: &str,
-    ts: Option<u64>,
-    query_id: Option<u64>,
-    empty_body: bool,
-) -> Result<impl IntoResponse, HttpError> {
-    let make_headers = |reader: &Arc<RwLock<RecordReader>>, last| {
-        let mut headers = HeaderMap::new();
-
-        let reader = reader.read().unwrap();
-        for (k, v) in reader.labels() {
-            headers.insert(
-                format!("x-reduct-label-{}", k)
-                    .parse::<HeaderName>()
-                    .unwrap(),
-                v.parse().unwrap(),
-            );
-        }
-
-        headers.insert(
-            "content-type",
-            reader.content_type().to_string().parse().unwrap(),
-        );
-        headers.insert(
-            "content-length",
-            reader.content_length().to_string().parse().unwrap(),
-        );
-        headers.insert(
-            "x-reduct-time",
-            reader.timestamp().to_string().parse().unwrap(),
-        );
-        headers.insert("x-reduct-last", u8::from(last).to_string().parse().unwrap());
-        headers
-    };
-
-    let (reader, last) = if let Some(ts) = ts {
-        let reader = bucket.begin_read(entry_name, ts)?;
-        (reader, true)
-    } else {
-        bucket.next(entry_name, query_id.unwrap())?
-    };
-
-    let headers = make_headers(&reader, last);
-
-    struct ReaderWrapper {
-        reader: Arc<RwLock<RecordReader>>,
-        empty_body: bool,
-    }
-
-    /// A wrapper around a `RecordReader` that implements `Stream` with RwLock
-    impl Stream for ReaderWrapper {
-        type Item = Result<Bytes, HttpError>;
-
-        fn poll_next(
-            self: Pin<&mut ReaderWrapper>,
-            _cx: &mut Context<'_>,
-        ) -> Poll<Option<Self::Item>> {
-            if self.empty_body {
-                return Poll::Ready(None);
-            }
-
-            if self.reader.read().unwrap().is_done() {
-                return Poll::Ready(None);
-            }
-            match self.reader.write().unwrap().read() {
-                Ok(chunk) => Poll::Ready(Some(Ok(chunk.unwrap()))),
-                Err(e) => Poll::Ready(Some(Err(e))),
-            }
-        }
-        fn size_hint(&self) -> (usize, Option<usize>) {
-            (0, None)
-        }
-    }
-
-    Ok((
-        headers,
-        StreamBody::new(ReaderWrapper { reader, empty_body }),
-    ))
-}
-
 impl IntoResponse for QueryInfo {
     fn into_response(self) -> Response {
         let mut headers = HeaderMap::new();
@@ -328,274 +130,19 @@ impl IntoResponse for QueryInfo {
     }
 }
 
-impl EntryApi {
-    // POST /:bucket/:entry?ts=<number>
-    pub async fn write_record(
-        State(components): State<Arc<RwLock<HttpServerState>>>,
-        headers: HeaderMap,
-        Path(path): Path<HashMap<String, String>>,
-        Query(params): Query<HashMap<String, String>>,
-        mut stream: BodyStream,
-    ) -> Result<(), HttpError> {
-        let bucket = path.get("bucket_name").unwrap();
-        check_permissions(
-            Arc::clone(&components),
-            headers.clone(),
-            WriteAccessPolicy {
-                bucket: bucket.clone(),
-            },
-        )?;
+impl EntryApi {}
 
-        let check_request_and_get_writer = || -> Result<Arc<RwLock<RecordWriter>>, HttpError> {
-            if !params.contains_key("ts") {
-                return Err(HttpError::unprocessable_entity(
-                    "'ts' parameter is required",
-                ));
-            }
-
-            let ts = match params.get("ts").unwrap().parse::<u64>() {
-                Ok(ts) => ts,
-                Err(_) => {
-                    return Err(HttpError::unprocessable_entity(
-                        "'ts' must be an unix timestamp in microseconds",
-                    ));
-                }
-            };
-            let content_size = headers
-                .get("content-length")
-                .ok_or(HttpError::unprocessable_entity(
-                    "content-length header is required",
-                ))?
-                .to_str()
-                .unwrap()
-                .parse::<u64>()
-                .map_err(|_| {
-                    HttpError::unprocessable_entity("content-length header must be a number")
-                })?;
-
-            let content_type = headers
-                .get("content-type")
-                .map_or("application/octet-stream", |v| v.to_str().unwrap())
-                .to_string();
-
-            let mut labels = Labels::new();
-            for (k, v) in headers.iter() {
-                if k.as_str().starts_with("x-reduct-label-") {
-                    let key = k.as_str()[15..].to_string();
-                    let value = match v.to_str() {
-                        Ok(value) => value.to_string(),
-                        Err(_) => {
-                            return Err(HttpError::unprocessable_entity(&format!(
-                                "Label values for {} must be valid UTF-8 strings",
-                                k
-                            )));
-                        }
-                    };
-                    labels.insert(key, value);
-                }
-            }
-
-            let writer = {
-                let mut components = components.write().unwrap();
-                let bucket = components.storage.get_bucket(bucket)?;
-                bucket.begin_write(
-                    path.get("entry_name").unwrap(),
-                    ts,
-                    content_size,
-                    content_type,
-                    labels,
-                )?
-            };
-            Ok(writer)
-        };
-
-        match check_request_and_get_writer() {
-            Ok(writer) => {
-                while let Some(chunk) = stream.next().await {
-                    let mut writer = writer.write().unwrap();
-                    let chunk = match chunk {
-                        Ok(chunk) => chunk,
-                        Err(e) => {
-                            writer.write(Chunk::Error)?;
-                            error!("Error while receiving data: {}", e);
-                            return Err(HttpError::from(e));
-                        }
-                    };
-                    writer.write(Chunk::Data(chunk))?;
-                }
-
-                writer.write().unwrap().write(Chunk::Last(Bytes::new()))?;
-                Ok(())
-            }
-            Err(e) => {
-                // drain the stream in the case if a client doesn't support Expect: 100-continue
-                if !headers.contains_key(Expect::name()) {
-                    debug!("draining the stream");
-                    while let Some(_) = stream.next().await {}
-                }
-
-                Err(e)
-            }
-        }
-    }
-
-    // GET /:bucket/:entry?ts=<number>|q=<number>|
-    pub async fn read_single_record(
-        State(components): State<Arc<RwLock<HttpServerState>>>,
-        Path(path): Path<HashMap<String, String>>,
-        Query(params): Query<HashMap<String, String>>,
-        headers: HeaderMap,
-        method: MethodExtractor,
-    ) -> Result<impl IntoResponse, HttpError> {
-        let bucket_name = path.get("bucket_name").unwrap();
-        let entry_name = path.get("entry_name").unwrap();
-        check_permissions(
-            Arc::clone(&components),
-            headers,
-            ReadAccessPolicy {
-                bucket: bucket_name.clone(),
-            },
-        )?;
-
-        let (query_id, ts) =
-            check_and_extract_ts_or_query_id(&components, params, bucket_name, entry_name)?;
-
-        fetch_and_response_single_record(
-            components
-                .write()
-                .unwrap()
-                .storage
-                .get_bucket(bucket_name)?,
-            entry_name,
-            ts,
-            query_id,
-            method.name() == "HEAD",
+pub fn create_entry_api_routes() -> axum::Router<Arc<RwLock<HttpServerState>>> {
+    axum::Router::new()
+        .route("/:bucket_name/:entry_name", post(write_record))
+        .route("/:bucket_name/:entry_name", get(read_single_record))
+        .route("/:bucket_name/:entry_name", head(read_single_record))
+        .route("/:bucket_name/:entry_name/batch", get(read_batched_records))
+        .route(
+            "/:bucket_name/:entry_name/batch",
+            head(read_batched_records),
         )
-    }
-
-    // GET /:bucket/:entry/batch?q=<number>
-    pub async fn read_batched_records(
-        State(components): State<Arc<RwLock<HttpServerState>>>,
-        Path(path): Path<HashMap<String, String>>,
-        Query(params): Query<HashMap<String, String>>,
-        headers: HeaderMap,
-        method: MethodExtractor,
-    ) -> Result<impl IntoResponse, HttpError> {
-        let bucket_name = path.get("bucket_name").unwrap();
-        let entry_name = path.get("entry_name").unwrap();
-        check_permissions(
-            Arc::clone(&components),
-            headers,
-            ReadAccessPolicy {
-                bucket: bucket_name.clone(),
-            },
-        )?;
-
-        let query_id = match params.get("q") {
-            Some(query) => query
-                .parse::<u64>()
-                .map_err(|_| HttpError::unprocessable_entity("'query' must be a number"))?,
-
-            None => {
-                return Err(HttpError::unprocessable_entity(
-                    "'q' parameter is required for batched reads",
-                ));
-            }
-        };
-
-        fetch_and_response_batched_records(
-            components
-                .write()
-                .unwrap()
-                .storage
-                .get_bucket(bucket_name)?,
-            entry_name,
-            query_id,
-            method.name == "HEAD",
-        )
-    }
-
-    // GET /:bucket/:entry/q?start=<number>&stop=<number>&continue=<number>&exclude-<label>=<value>&include-<label>=<value>&ttl=<number>
-    pub async fn query(
-        State(components): State<Arc<RwLock<HttpServerState>>>,
-        Path(path): Path<HashMap<String, String>>,
-        Query(params): Query<HashMap<String, String>>,
-        headers: HeaderMap,
-    ) -> Result<QueryInfo, HttpError> {
-        let bucket_name = path.get("bucket_name").unwrap();
-        let entry_name = path.get("entry_name").unwrap();
-
-        check_permissions(
-            Arc::clone(&components),
-            headers,
-            ReadAccessPolicy {
-                bucket: bucket_name.clone(),
-            },
-        )?;
-
-        let entry_info = {
-            let mut components = components.write().unwrap();
-            let bucket = components.storage.get_bucket(bucket_name)?;
-            bucket.get_entry(entry_name)?.info()?
-        };
-
-        let start = match params.get("start") {
-            Some(start) => start.parse::<u64>().map_err(|_| {
-                HttpError::unprocessable_entity("'start' must be an unix timestamp in microseconds")
-            })?,
-            None => entry_info.oldest_record,
-        };
-
-        let stop = match params.get("stop") {
-            Some(stop) => stop.parse::<u64>().map_err(|_| {
-                HttpError::unprocessable_entity("'stop' must be an unix timestamp in microseconds")
-            })?,
-            None => entry_info.latest_record + 1,
-        };
-
-        let continuous = match params.get("continuous") {
-            Some(continue_) => continue_.parse::<bool>().map_err(|_| {
-                HttpError::unprocessable_entity(
-                    "'continue' must be an unix timestamp in microseconds",
-                )
-            })?,
-            None => false,
-        };
-
-        let ttl = match params.get("ttl") {
-            Some(ttl) => ttl.parse::<u64>().map_err(|_| {
-                HttpError::unprocessable_entity("'ttl' must be an unix timestamp in microseconds")
-            })?,
-            None => 5,
-        };
-
-        let mut include = HashMap::new();
-        let mut exclude = HashMap::new();
-
-        for (k, v) in params.iter() {
-            if k.starts_with("include-") {
-                include.insert(k[8..].to_string(), v.to_string());
-            } else if k.starts_with("exclude-") {
-                exclude.insert(k[8..].to_string(), v.to_string());
-            }
-        }
-
-        let mut components = components.write().unwrap();
-        let bucket = components.storage.get_bucket(bucket_name)?;
-        let entry = bucket.get_or_create_entry(entry_name)?;
-        let id = entry.query(
-            start,
-            stop,
-            QueryOptions {
-                continuous,
-                include,
-                exclude,
-                ttl: Duration::from_secs(ttl),
-            },
-        )?;
-
-        Ok(QueryInfo { id })
-    }
+        .route("/:bucket_name/:entry_name/q", get(query::query))
 }
 
 #[cfg(test)]
@@ -614,7 +161,7 @@ mod tests {
     use std::path::PathBuf;
 
     #[fixture]
-    fn components() -> Arc<RwLock<HttpServerState>> {
+    pub fn components() -> Arc<RwLock<HttpServerState>> {
         let data_path = tempfile::tempdir().unwrap().into_path();
 
         let mut components = HttpServerState {
@@ -643,7 +190,7 @@ mod tests {
     }
 
     #[fixture]
-    fn headers() -> HeaderMap {
+    pub fn headers() -> HeaderMap {
         let mut headers = HeaderMap::new();
         headers.insert("content-length", "0".parse().unwrap());
         headers.insert("x-reduct-label-x", "y".parse().unwrap());
@@ -651,183 +198,11 @@ mod tests {
     }
 
     #[fixture]
-    fn path() -> Path<HashMap<String, String>> {
+    pub fn path() -> Path<HashMap<String, String>> {
         let path = Path(HashMap::from_iter(vec![
             ("bucket_name".to_string(), "bucket-1".to_string()),
             ("entry_name".to_string(), "entry-1".to_string()),
         ]));
         path
-    }
-
-    #[rstest]
-    #[tokio::test]
-    async fn test_write_with_label_ok(
-        components: Arc<RwLock<HttpServerState>>,
-        headers: HeaderMap,
-        path: Path<HashMap<String, String>>,
-    ) {
-        let emtpy_stream: Empty<Bytes> = Empty::new();
-        let request = Request::builder().body(emtpy_stream).unwrap();
-        let body = BodyStream::from_request(request, &()).await.unwrap();
-
-        EntryApi::write_record(
-            State(Arc::clone(&components)),
-            headers,
-            path,
-            Query(HashMap::from_iter(vec![(
-                "ts".to_string(),
-                "1".to_string(),
-            )])),
-            body,
-        )
-        .await
-        .unwrap();
-
-        let record = components
-            .write()
-            .unwrap()
-            .storage
-            .get_bucket("bucket-1")
-            .unwrap()
-            .begin_read("entry-1", 1)
-            .unwrap();
-
-        assert_eq!(
-            record.read().unwrap().labels().get("x"),
-            Some(&"y".to_string())
-        );
-    }
-
-    #[rstest]
-    #[case("GET", "Hey!!!")]
-    #[case("HEAD", "")]
-    #[tokio::test]
-    async fn test_single_read_ts(
-        components: Arc<RwLock<HttpServerState>>,
-        path: Path<HashMap<String, String>>,
-        #[case] method: String,
-        #[case] body: String,
-    ) {
-        let mut response = EntryApi::read_single_record(
-            State(Arc::clone(&components)),
-            path,
-            Query(HashMap::from_iter(vec![(
-                "ts".to_string(),
-                "0".to_string(),
-            )])),
-            HeaderMap::new(),
-            MethodExtractor::new(&method),
-        )
-        .await
-        .unwrap()
-        .into_response();
-
-        let headers = response.headers();
-        assert_eq!(headers["x-reduct-time"], "0");
-        assert_eq!(headers["content-type"], "text/plain");
-        assert_eq!(headers["content-length"], "6");
-
-        assert_eq!(
-            response.data().await.unwrap_or(Ok(Bytes::new())).unwrap(),
-            Bytes::from(body)
-        );
-    }
-
-    #[rstest]
-    #[case("GET", "Hey!!!")]
-    #[case("HEAD", "")]
-    #[tokio::test]
-    async fn test_single_read_query(
-        components: Arc<RwLock<HttpServerState>>,
-        path: Path<HashMap<String, String>>,
-        #[case] method: String,
-        #[case] body: String,
-    ) {
-        let query_id = query_records(&components);
-
-        let mut response = EntryApi::read_single_record(
-            State(Arc::clone(&components)),
-            path,
-            Query(HashMap::from_iter(vec![(
-                "q".to_string(),
-                query_id.to_string(),
-            )])),
-            HeaderMap::new(),
-            MethodExtractor::new(&method),
-        )
-        .await
-        .unwrap()
-        .into_response();
-
-        let headers = response.headers();
-        assert_eq!(headers["x-reduct-time"], "0");
-        assert_eq!(headers["content-type"], "text/plain");
-        assert_eq!(headers["content-length"], "6");
-
-        assert_eq!(
-            response.data().await.unwrap_or(Ok(Bytes::new())).unwrap(),
-            Bytes::from(body)
-        );
-    }
-
-    fn query_records(components: &Arc<RwLock<HttpServerState>>) -> u64 {
-        let query_id = components
-            .write()
-            .unwrap()
-            .storage
-            .get_bucket("bucket-1")
-            .unwrap()
-            .get_entry("entry-1")
-            .unwrap()
-            .query(
-                0,
-                1,
-                QueryOptions {
-                    continuous: false,
-                    include: HashMap::new(),
-                    exclude: HashMap::new(),
-                    ttl: Duration::from_secs(1),
-                },
-            )
-            .unwrap();
-        query_id
-    }
-
-    #[rstest]
-    #[case("GET", "Hey!!!")]
-    #[case("HEAD", "")]
-    #[tokio::test]
-    async fn test_batched_read(
-        components: Arc<RwLock<HttpServerState>>,
-        path: Path<HashMap<String, String>>,
-        #[case] method: String,
-        #[case] body: String,
-    ) {
-        let query_id = query_records(&components);
-
-        let mut response = EntryApi::read_batched_records(
-            State(Arc::clone(&components)),
-            path,
-            Query(HashMap::from_iter(vec![(
-                "q".to_string(),
-                query_id.to_string(),
-            )])),
-            HeaderMap::new(),
-            MethodExtractor::new(method.as_str()),
-        )
-        .await
-        .unwrap()
-        .into_response();
-
-        let headers = response.headers();
-        assert_eq!(headers["x-reduct-time-0"], "6,text/plain,b=\"[a,b]\",x=y");
-        assert_eq!(headers["content-type"], "application/octet-stream");
-        assert_eq!(headers["content-length"], "6");
-        assert_eq!(headers["x-reduct-last"], "true");
-
-        assert_eq!(
-            response.data().await.unwrap_or(Ok(Bytes::new())).unwrap(),
-            Bytes::from(body)
-        );
     }
 }
